Reworked UnityContent framework
