+++ source/Exports.ts
@@ -1,5 +1,5 @@
 import Unity from "./components/Unity";
-import UnityContent from "./UnityContent";
+import UnityContent from "./models/UnityContent";
 
 export default Unity;
 export { UnityContent };
